create loader for inifinite scroll
