+++ packages/apps/docs/src/components/InfiniteScroll/InfiniteScroll.tsx
@@ -1,0 +1,12 @@
+import { Loading } from '@/components';
+import React, { FC } from 'react';
+
+interface IProps {}
+
+export const InfiniteScroll: FC<IProps> = () => {
+  return (
+    <div>
+      <Loading />
+    </div>
+  );
+};
