+++ packages/apps/docs/src/components/InfiniteScroll/index.ts
@@ -1,0 +1,1 @@
+export { InfiniteScroll } from './InfiniteScroll';
