+++ packages/apps/docs/src/components/Search/styles.ts
@@ -1,4 +1,4 @@
-import { keyframes, styled, StyledComponent } from '@kadena/react-components';
+import { styled, StyledComponent } from '@kadena/react-components';
 
 export const SearchForm: StyledComponent<'form'> = styled('form', {
   width: '100%',
@@ -60,15 +60,4 @@ export const LoadingWrapper: StyledComponent<'div'> = styled('div', {
   justifyContent: 'center',
   padding: '$10 0',
   background: '$backgroundOverlayColor',
-});
-
-const rotate = keyframes({
-  '0%': { transform: 'rotate(0deg)' },
-  '100%': { transform: 'rotate(360deg)' },
-});
-
-export const Loader: StyledComponent<'div'> = styled('div', {
-  width: '$6',
-  height: '$6',
-  animation: `${rotate} 1s infinite linear `,
 });
