+++ packages/apps/docs/src/components/index.ts
@@ -10,3 +10,5 @@ export * from './LastModifiedDate';
 export * from './Search';
 export * from './SearchModal';
 export * from './MostPopular';
+export * from './InfiniteScroll';
+export * from './Loading';
